+++ src/components/api/accessory.types.ts
@@ -1,0 +1,7 @@
+export interface AccessoryData {
+  id: number;
+  name: string;
+  price: number;
+  imageUrl: string;
+  explanation: string;
+}
