+++ src/components/api/pattern.ts
@@ -1,8 +1,8 @@
 import axios, { AxiosResponse } from 'axios';
-import { WheelData } from './wheels.types';
+import { PatternData } from './pattern.types';
 
 const instance = axios.create({
-  baseURL: `${process.env.REACT_APP_API_URL}/api/wheels`,
+  baseURL: `${process.env.REACT_APP_API_URL}/api/patterns`,
   timeout: 15000,
   withCredentials: true,
 });
@@ -10,10 +10,10 @@ const instance = axios.create({
 const responseBody = (response: AxiosResponse) => response.data;
 // const responseStatus = (response: AxiosResponse) => response.status;
 
-const wheelRequests = {
+const patternRequests = {
   get: <T>(url: string) => instance.get<T>(url).then(responseBody),
 };
 
-export const Wheel = {
-  list: (): Promise<WheelData[]> => wheelRequests.get<WheelData[]>(''),
+export const Pattern = {
+  list: (): Promise<PatternData[]> => patternRequests.get<PatternData[]>(''),
 };
