+++ src/components/api/wheel.ts
@@ -1,8 +1,8 @@
 import axios, { AxiosResponse } from 'axios';
-import { ColorData } from './colors.types';
+import { WheelData } from './wheel.types';
 
 const instance = axios.create({
-  baseURL: `${process.env.REACT_APP_API_URL}/api/colors`,
+  baseURL: `${process.env.REACT_APP_API_URL}/api/wheels`,
   timeout: 15000,
   withCredentials: true,
 });
@@ -10,10 +10,10 @@ const instance = axios.create({
 const responseBody = (response: AxiosResponse) => response.data;
 // const responseStatus = (response: AxiosResponse) => response.status;
 
-const colorRequests = {
+const wheelRequests = {
   get: <T>(url: string) => instance.get<T>(url).then(responseBody),
 };
 
-export const Color = {
-  list: (): Promise<ColorData[]> => colorRequests.get<ColorData[]>(''),
+export const Wheel = {
+  list: (): Promise<WheelData[]> => wheelRequests.get<WheelData[]>(''),
 };
